Rename SLURM attributes
